+++ index.js
@@ -17,6 +17,7 @@ const Layer = require('./lib/layer')
 const { METHODS } = require('node:http')
 const parseUrl = require('parseurl')
 const Route = require('./lib/route')
+const debug = require('debug')('router')
 
 /**
  * Module variables.
@@ -148,6 +149,8 @@ Router.prototype.handle = function handle (req, res, callback) {
   if (!callback) {
     throw new TypeError('argument callback is required')
   }
+
+  debug('dispatching %s %s', req.method, req.url)
 
   let idx = 0
   let methods
@@ -315,6 +318,7 @@ Router.prototype.handle = function handle (req, res, callback) {
 
       // Trim off the part of the url that matches the route
       // middleware (.use stuff) needs to have the path stripped
+      debug('trim prefix (%s) from url %s', layerPath, req.url)
       removed = layerPath
       req.url = protohost + req.url.slice(protohost.length + removed.length)
 
@@ -329,6 +333,8 @@ Router.prototype.handle = function handle (req, res, callback) {
         ? removed.substring(0, removed.length - 1)
         : removed)
     }
+
+    debug('%s %s : %s', layer.name, layerPath, req.originalUrl)
 
     if (layerError) {
       layer.handleError(layerError, req, res, next)
@@ -387,6 +393,8 @@ Router.prototype.use = function use (handler) {
     }
 
     // add the middleware
+    debug('use %o %s', path, fn.name || '<anonymous>')
+
     const layer = new Layer(path, {
       sensitive: this.caseSensitive,
       strict: false,