+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==================
+
+* Restore `debug` dependency
+
 2.1.0 / 2025-02-10
 ==================